+++ Examples/selection.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-# Auto-Generated .sh files from the original .md by Sparkle 0.9.0
+# Auto-Generated .sh files from the original .md by Sparkle 0.9.2
 
 ## Algorithm Selection
 
@@ -35,7 +35,7 @@ sparkle add feature extractor Examples/Resources/Extractors/SAT-features-competi
 sparkle compute features
 
 ### Run the solvers
-# Similarly, we can now also compute our objective values for our solvers, in this case PAR10. Note that we can at this point still specifiy multiple objectives by separating them with a comma, or denote them in our settings file.
+# Similarly, we can now also compute our objective values for our solvers, in this case PAR10 as specified in the settings file. We run the run solvers command with the `--performance-data`, so Sparkle will compute all empty values in the performance data frame.
 
 sparkle run solvers --performance-data
 
@@ -44,7 +44,7 @@ sparkle run solvers --performance-data
 
 sparkle jobs
 
-# Now we can construct a portfolio selector, using the previously computed features and the results of running the solvers. The `--selector-timeout` argument determines for how many seconds we will train our selector for. We can set the flag `--solver-ablation` for actual marginal contribution computation later.
+# Now we can construct a portfolio selector, using the previously computed features and the objective value results of running the solvers. We can specify an objective to select on with the `--objective` flag, but if we do not, Sparkle will default to the first objective specified in the Settings file. The `--selector-timeout` argument determines for how many seconds we will train our selector for. We can set the flag `--solver-ablation` for actual marginal contribution computation later.
 
 sparkle construct portfolio selector --selector-timeout 1000 --solver-ablation
 sparkle jobs  # Wait for the constructor to complete its computations
@@ -101,13 +101,13 @@ sparkle jobs  # Wait for it to complete before continuing
 
 # And run the solvers on the new data set.
 
-sparkle run solvers
+sparkle run solvers --performance-data
 sparkle jobs
 
 # Now we can train a selector based on these features.
 
 sparkle construct portfolio selector --selector-timeout 1000
-sparkle jobs  #Wait for the computation to be done
+sparkle jobs  # Wait for the computation to be done
 
 # And generate the report. When running on the PTN/PTN2 data sets, you can compare the two to see the impact of different feature extractors.